added trading

+++ MarketSimulator.Server/Entities/Asset.cs
@@ -7,6 +7,6 @@
         public Guid CompanyId;
         public IndustryTypes PrimaryIndustry;
         public IndustryTypes SecondaryIndustry;
-        public decimal Value;
+        public double Value;
     }
 }
